Add the Logic to Handle "Others" Purchase Type (#59)
Say we want to purchase something at Canadian Tire, but none of our
credit cards has a home improvement purchase reward. In this situation,
the app should use "Others" purchase reward to determine the reward
amount. This PR adds this logic and a safeguard to ensure there won't be
a null reference exception.